feat: add CORS configuration for GET requests
This commit adds a CORS configuration to allow GET requests from any origin. This is necessary to enable cross-origin resource sharing for the GET requests made to the server. By specifying the origin as "*", we allow requests from any domain. This change improves the flexibility and accessibility of the API.